+++ src/components/Footer/styles/footer.js
@@ -1,0 +1,27 @@
+
+export const myStyle = {
+    backgroundColor: '#424a4d',
+    display: 'flex',
+    justifyContent: 'flex-start',
+    alignItems: 'flex-start',
+    padding: '30px',
+  };
+  
+  export const myStyle2 = {
+    listStyleType: 'none',
+    color: '#f1eee4',
+  };
+  
+  export const myStyle3 = {
+    textDecoration: 'none',
+    color: '#f1eee4',
+    padding: '20px',
+  };
+  
+  export const myStyle4 = {
+    margin: '0',
+    backgroundColor: '#424a4d',
+    color: '#f1eee4',
+    textAlign: 'center'
+  }
+
