+++ src/components/Header/styles/header.js
@@ -1,0 +1,6 @@
+
+export const myStyle = {
+    backgroundColor: "#424a4d",
+    color: "#f1eee4",
+    paddingRight: "50px"
+  }
