+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { default as Header } from './Header/Index'
+export { default as Footer } from './Footer/Index'
