+++ src/pages/Create-NFT/index.jsx
@@ -1,27 +1,7 @@
 import React from "react";
 import { Form, Button, FloatingLabel } from "react-bootstrap";
+import { pageTemplate, form, h1 } from './styles/createNFT'
 
-const pageTemplate = {
-  display: "flex",
-  flexDirection: "row",
-  backgroundColor: "#f1eee4",
-  color: "#424a4d",
-  width: "100%",
-  height: "100%",
-  margin: "0",
-  justifyContent: "center",
-};
-
-const form = {
-  width: "50%",
-  padding: "30px",
-  marginBottom: "60px"
-};
-
-const h1 = {
-  textAlign: "center",
-  marginBottom: "60px",
-};
 
 export default function CreateNFT() {
   return (
