+++ src/pages/Create-NFT/styles/createNFT.js
@@ -1,0 +1,21 @@
+export const pageTemplate = {
+    display: "flex",
+    flexDirection: "row",
+    backgroundColor: "#f1eee4",
+    color: "#424a4d",
+    width: "100%",
+    height: "100%",
+    margin: "0",
+    justifyContent: "center",
+  };
+  
+ export  const form = {
+    width: "50%",
+    padding: "30px",
+    marginBottom: "60px"
+  };
+  
+ export  const h1 = {
+    textAlign: "center",
+    marginBottom: "60px",
+  };
