--- src/pages/Gallery.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-
-const pageTemplate = {
-  backgroundColor: "#f1eee4",
-  color: "#424a4d",
-  width: "100%",
-  height: "515px",
-  padding: "50px",
-  margin: "50px",
-};
-
-export default function Gallery() {
-  return <div style={pageTemplate}></div>;
-}
