+++ src/pages/Gallery/index.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { pageTemplate } from './styles/gallery'
+
+export default function Gallery() {
+  return <div style={pageTemplate}></div>;
+}
