+++ src/pages/Gallery/styles/gallery.js
@@ -1,0 +1,9 @@
+export const pageTemplate = {
+    backgroundColor: "#f1eee4",
+    color: "#424a4d",
+    width: "100%",
+    height: "515px",
+    padding: "50px",
+    margin: "50px",
+  };
+
