+++ src/pages/Home/index.jsx
@@ -1,27 +1,6 @@
 import React from 'react';
 import logo from "../assets/home_logo.png"
-
-const pageTemplate = {
-  backgroundColor: "#f1eee4",
-  color: "#424a4d",
-  width: "100%",
-  height: "515px",
-  padding: "50px",
-  margin: "50px"
-}
-
-const flex = {
-  display: "flex",
-  justifyContent: "center",
-  backgroundColor: "#f1eee4",
-}
-
-const logoImage = {
-  width: "500px",
-  height: "500px",
-  padding: "20px"
-}
-
+import { pageTemplate, flex, logoImage } from './styles/Home';
 
 export default function Home() {
   return (
