+++ src/pages/Home/styles/Home.js
@@ -1,0 +1,21 @@
+export const pageTemplate = {
+    backgroundColor: "#f1eee4",
+    color: "#424a4d",
+    width: "100%",
+    height: "515px",
+    padding: "50px",
+    margin: "50px"
+  }
+  
+  export const flex = {
+    display: "flex",
+    justifyContent: "center",
+    backgroundColor: "#f1eee4",
+  }
+  
+  export const logoImage = {
+    width: "500px",
+    height: "500px",
+    padding: "20px"
+  }
+
