+++ src/pages/index.js
@@ -1,0 +1,3 @@
+export { default as Home } from './Header/Index'
+export { default as Gallery } from './Gallery/index'
+export { default as CreateNFT } from './Create-NFT/index'
